Reuse paz.transformers across Whisper; flatten and abstract
Address review: lift Whisper's reusable machinery into shared paz modules and
have the foundation model consume them, instead of reimplementing.

- paz.transformers.attention: generic EinsumDense multi-head attention
  (attend, kv_attend, build/update_cache, projections) extracted from the
  Whisper layer; kv_attend takes explicit params (no config bundle).
- paz.transformers.embeddings.absolute: learned positional embeddings.
- paz.transformers.tokenizers: byte-level BPE decode core (GPT-2 family).
- paz.audio: log-mel frontend (STFT + mel filterbank) moved into the audio
  backend so other models (e.g. Gemma audio) can reuse it.

Whisper now imports all of the above; its layers/ keeps only the model-specific
encoder/decoder block assembly. Builders live in a flatter model.py (no
models/whisper.py) and Whisper(...) returns a namedtuple
Whisper(frontend, encoder, cross_cache, decoder_step). gemma4 attention is left
as-is for now (RoPE/GQA/soft-cap differ); it can adopt the shared core later.

Co-located tests added for attention, tokenizers, absolute embeddings and the
audio frontend. The harvard.wav token-id snapshot still matches bit-for-bit,
confirming the move is behavior-preserving (weights load by identical names).
