+++ paz/backend/audio_test.py
@@ -1,6 +1,8 @@
 import numpy as np
+from keras import ops
 
 from paz.backend.audio import to_float, to_mono, resample
+from paz.backend.audio import build_mel_filters, log_mel_spectrogram
 
 
 def test_to_float_int16_full_scale():
@@ -38,3 +40,11 @@ def test_resample_keeps_rate_when_equal():
     waveform = np.linspace(-1.0, 1.0, 100, dtype="float32")
     result = resample(waveform, 16000, 16000)
     np.testing.assert_allclose(result, waveform)
+
+
+def test_log_mel_spectrogram_has_mel_channels():
+    mel_filters = build_mel_filters(80, 400, 16000, 45.245640471924965)
+    mel_filters = ops.convert_to_tensor(mel_filters)
+    waveform = ops.zeros((1, 16000), dtype="float32")
+    features = log_mel_spectrogram(waveform, mel_filters, 16000 * 30, 400, 160)
+    assert features.shape[-1] == 80
