+++ paz/models/transformers/embeddings/__init__.py
@@ -1,1 +1,2 @@
 from paz.models.transformers.embeddings import rotary
+from paz.models.transformers.embeddings import absolute
